Implement LinkEquipment.Factory
LinkEquipment.Factory is a general entrypoint into entity factories via
build(<entity_module>) and insert(<entity_module>) in tests. It further
provides a behavior for entity factories to implement and offers a
default implementation of insert.

In the future it would be nice to add the verification of the provided
fields via the __using__ macro of LinkEquipment.Factory.